+++ packages/app-config-writer/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @sap-ux/app-config-writer
 
+## 0.6.78
+
+### Patch Changes
+
+- Updated dependencies [55ac9f0]
+    - @sap-ux/project-access@1.34.1
+    - @sap-ux/axios-extension@1.25.6
+
+## 0.6.77
+
+### Patch Changes
+
+- Updated dependencies [b132944]
+    - @sap-ux/project-access@1.34.0
+    - @sap-ux/axios-extension@1.25.6
+
 ## 0.6.76
 
 ### Patch Changes