+++ packages/backend-proxy-middleware-cf/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @sap-ux/backend-proxy-middleware-cf
 
+## 0.0.28
+
+### Patch Changes
+
+- Updated dependencies [55ac9f0]
+    - @sap-ux/project-access@1.34.1
+    - @sap-ux/adp-tooling@0.18.38
+
+## 0.0.27
+
+### Patch Changes
+
+- Updated dependencies [b132944]
+    - @sap-ux/project-access@1.34.0
+    - @sap-ux/adp-tooling@0.18.37
+
+## 0.0.26
+
+### Patch Changes
+
+- ce4b29c: Upgrade qs/body-parser/express
+- Updated dependencies [ce4b29c]
+    - @sap-ux/adp-tooling@0.18.36
+
 ## 0.0.25
 
 ### Patch Changes